requirements-test.txt
# Base requirements
-r requirements.txt

# Test framework
coverage
pytest
pytest-django
